fix(ci): python-coverage-shards - split oversized coverage tests
